Refactor and fix stub validation errors
