+++ metadata-ingestion/tests/unit/glue/glue_mces_golden.json
@@ -525,7 +525,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -562,7 +563,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -599,7 +601,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -637,7 +640,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -674,7 +678,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -709,7 +714,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -786,7 +792,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -823,7 +830,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -860,7 +868,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
@@ -897,7 +906,8 @@
               "type": {
                 "com.linkedin.pegasus2avro.datajob.azkaban.AzkabanJobType": "GLUE"
               },
-              "flowUrn": null
+              "flowUrn": null,
+              "status": null
             }
           },
           {
